requirements.txt
streamlit
pandas
pypdf
langchain-google-genai
python-dotenv
langchain-core
